LLM with RAG/pyproject.toml
[project]
name = "llm-with-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.8.1",
    "bitsandbytes>=0.45.3",
    "chromadb>=1.0.15",
    "editdistance>=0.8.1",
    "gradio>=5.35.0",
    "jiwer>=4.0.0",
    "langchain-community>=0.3.21",
    "lhotse>=1.30.3",
    "librosa>=0.11.0",
    "nemo-toolkit[nlp]>=2.3.1",
    "notebook>=7.4.4",
    "openai>=1.93.0",
    "openai-whisper>=20250625",
    "pvorca>=1.2.0",
    "pyannote-core>=5.0.0",
    "pyannote-metrics>=3.2.1",
    "sounddevice>=0.5.2",
    "spaces>=0.26.0",
    "torchaudio>=2.7.1",
    "transformers>=4.51.3",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
